requirements.txt
django
djangorestframework
django-assets
pytest
pytest-django
pytest-cov
closure
yuicompressor
factory_boy
psycopg2
